projeto_rpa/requirements.txt
playwright
pandas
gspread
oauth2client
python-telegram-bot
python-dotenv